Create Web_Scrapping(JavaScript-Ajax (Dynamic) Websites).ipynb
Scrape data that loaded by ajax requests and one of the solutions to do it is to use a selenium webdriver but the selenium is heavy thin and the scrape with it is a rather slow process so I would scrape ajax website with the only requests library and beautifulsoup library. I used the steam store as an example.